CAS1 SAR Template/Data Tidy Up
Remove internal identifiers
Remove fields that are duplicated across the SAR
Remove data fields that are not used in the template
Ensure we populate all fields in the sample report generated by tests
Fix field mappings as required
