+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/jpa/entity/SubjectAccessRequestRepositoryBase.kt
@@ -8,224 +8,6 @@ import java.time.LocalDateTime
 @Repository
 open class SubjectAccessRequestRepositoryBase(val jdbcTemplate: NamedParameterJdbcTemplate) {
 
-  fun bookings(
-    crn: String?,
-    nomsNumber: String?,
-    startDate: LocalDateTime?,
-    endDate: LocalDateTime?,
-  ): String? {
-    val result = jdbcTemplate.queryForMap(
-      """
-    select json_agg(booking) as json 
-    from (
-          select
-               b.crn ,
-              b.noms_number,
-              b.arrival_date,
-              b.departure_date,
-              b.original_arrival_date,
-              b.original_departure_date,
-              b.created_at,
-              b.status,
-              p."name" as premises_name,
-              b.adhoc,
-              b.key_worker_staff_code,
-              b.offender_name
-          from
-              cas3_bookings b
-          left join cas3_premises p on
-              b.premises_id = p.id
-          where
-              (b.crn = :crn
-              or b.noms_number = :noms_number )
-          and (:start_date::date is null or b.created_at >= :start_date) 
-          and (:end_date::date is null or b.created_at <= :end_date)
-  ) booking
-      """.trimIndent(),
-      MapSqlParameterSource().addSarParameters(
-        crn,
-        nomsNumber,
-        startDate,
-        endDate,
-      ),
-    )
-    return toJsonString(result)
-  }
-
-  fun spaceBookings(
-    crn: String?,
-    nomsNumber: String?,
-    startDate: LocalDateTime?,
-    endDate: LocalDateTime?,
-  ): String? {
-    val result = jdbcTemplate.queryForMap(
-      """
-    select json_agg(spaceBooking) as json 
-    from (
-          select
-            b.crn,
-            a.noms_number,            
-            b.canonical_arrival_date,
-            b.canonical_departure_date,
-            b.expected_arrival_date,
-            b.expected_departure_date,
-            b.actual_arrival_date,
-            b.actual_arrival_time,
-            b.actual_departure_date,
-            b.actual_departure_time,
-            b.non_arrival_confirmed_at,
-            b.non_arrival_notes,
-            nar.name as non_arrival_reason,
-            apa.risk_ratings -> 'tier' -> 'value' ->> 'level' as tier,
-            b.created_at,
-            b.key_worker_staff_code,
-            b.key_worker_assigned_at,
-            b.key_worker_name,
-            p."name" as premises_name,
-            b.delius_event_number,
-            u.name as created_by_user_name,
-            dr.name as departure_reason,
-            b.departure_notes,
-            moc.name as move_on_category,
-            b.cancellation_reason_notes,
-            cr.name as cancellation_reason,
-            b.cancellation_occurred_at, 
-            b.cancellation_recorded_at,
-            b.transfer_type,
-            b.additional_information,
-            b.transfer_reason,
-            ( 
-              SELECT STRING_AGG (cas1_characteristics.property_name, ',')
-              FROM cas1_space_bookings_criteria sbc
-              LEFT OUTER JOIN cas1_characteristics ON cas1_characteristics.id = sbc.characteristic_id
-              WHERE sbc.space_booking_id = b.id 
-              GROUP by sbc.space_booking_id
-            ) AS characteristics_property_names,    
-            CASE 
-              WHEN apa.id IS NOT NULL THEN apa.name
-              ELSE offline_app.name
-            END as person_name
-            FROM 
-              cas1_space_bookings b
-            LEFT JOIN non_arrival_reasons nar ON 
-              b.non_arrival_reason_id = nar.id           
-            LEFT JOIN cas1_premises_base p ON
-              b.premises_id = p.id            
-            LEFT OUTER JOIN approved_premises_applications apa ON 
-              b.approved_premises_application_id = apa.id
-            LEFT OUTER JOIN offline_applications offline_app ON 
-            b.offline_application_id = offline_app.id              
-            LEFT OUTER JOIN   
-              applications a on 
-              a.id = apa.id
-            LEFT JOIN departure_reasons dr ON 
-              b.departure_reason_id = dr.id
-            LEFT JOIN move_on_categories moc ON
-              b.departure_move_on_category_id = moc.id
-            LEFT JOIN cancellation_reasons cr ON
-              b.cancellation_reason_id = cr.id
-            LEFT JOIN users u ON
-              b.created_by_user_id = u.id
-          where
-              (b.crn = :crn
-              or a.noms_number = :noms_number )
-          and (:start_date::date is null or b.created_at >= :start_date) 
-          and (:end_date::date is null or b.created_at <= :end_date)         
-  ) spaceBooking
-      """.trimIndent(),
-      MapSqlParameterSource().addSarParameters(
-        crn,
-        nomsNumber,
-        startDate,
-        endDate,
-      ),
-    )
-    return toJsonString(result)
-  }
-
-  fun bookingExtensions(
-    crn: String?,
-    nomsNumber: String?,
-    startDate: LocalDateTime?,
-    endDate: LocalDateTime?,
-  ): String? {
-    var result = jdbcTemplate.queryForMap(
-      """
-        select json_agg(booking_ext) as json 
-        from (
-      select
-            b.crn,
-            b.noms_number,
-            e.previous_departure_date,
-            e.new_departure_date,
-            e.notes,
-            e.created_at
-        from
-            cas3_extensions e
-        join cas3_bookings b on
-            b.id = e.booking_id
-        left join applications a on
-            a.id = b.application_id
-        left join offline_applications oa on
-            oa.id = b.offline_application_id
-      where
-        (b.crn = :crn
-          or b.noms_number = :noms_number )
-      and (:start_date::date is null or b.created_at >= :start_date)
-      and (:end_date::date is null or b.created_at <= :end_date)
-        )booking_ext
-      """.trimIndent(),
-      MapSqlParameterSource().addSarParameters(
-        crn,
-        nomsNumber,
-        startDate,
-        endDate,
-      ),
-    )
-    return toJsonString(result)
-  }
-
-  fun cancellations(
-    crn: String?,
-    nomsNumber: String?,
-    startDate: LocalDateTime?,
-    endDate: LocalDateTime?,
-  ): String? {
-    var result = jdbcTemplate.queryForMap(
-      """
-          select json_agg(cancellation) as json
-          from (
-              select
-                  b.crn,
-                  b.noms_number,
-                  c.notes,
-                  c."date" as cancellation_date,
-                  cr."name" as cancellation_reason,
-                  c.other_reason,
-                  c.created_at
-              from
-                  cas3_cancellations c
-                  inner join cas3_bookings b on
-                      b.id = c.booking_id
-                  inner join cancellation_reasons cr on
-                      c.cancellation_reason_id = cr.id
-              where
-                  (b.crn = :crn
-                      or b.noms_number = :noms_number )
-                and (:start_date::date is null or b.created_at >= :start_date)
-                and (:end_date::date is null or b.created_at <= :end_date)        
-          ) cancellation
-      """.trimIndent(),
-      MapSqlParameterSource().addSarParameters(
-        crn,
-        nomsNumber,
-        startDate,
-        endDate,
-      ),
-    )
-    return toJsonString(result)
-  }
-
   fun domainEvents(
     crn: String?,
     nomsNumber: String?,
@@ -237,12 +19,10 @@ open class SubjectAccessRequestRepositoryBase(val jdbcTemplate: NamedParameterJd
       """
            select json_agg(domain_events) as json from ( 
                select 
-                 de.crn,
                  de."type",
                  de.occurred_at,
                  de.created_at,
                  de."data",
-                 de.noms_number,
                  CASE
                    WHEN cas1_3_app.id IS NOT NULL THEN cas1_3_app.submitted_at
                    WHEN cas2_app.id IS NOT NULL THEN cas2_app.submitted_at
@@ -267,42 +47,6 @@ open class SubjectAccessRequestRepositoryBase(val jdbcTemplate: NamedParameterJd
                and (:end_date::date is null or de.created_at <= :end_date) 
                order by de.created_at
            ) domain_events
-      """.trimIndent(),
-      MapSqlParameterSource()
-        .addSarParameters(crn, nomsNumber, startDate, endDate)
-        .addValue("service_name", serviceName),
-    )
-    return toJsonString(result)
-  }
-
-  fun domainEventMetadata(
-    crn: String?,
-    nomsNumber: String?,
-    startDate: LocalDateTime?,
-    endDate: LocalDateTime?,
-    serviceName: String = "CAS1",
-  ): String? {
-    val result = jdbcTemplate.queryForMap(
-      """
-             select json_agg(domain_events_metadata) as json
-             from ( 
-                 select 
-                     de.crn,
-                     de.noms_number,
-                     de.created_at,
-                     dem."name",
-                     dem.value
-                 from 
-                     domain_events_metadata dem 
-                 inner join domain_events de on 
-                     de.id = dem.domain_event_id
-                 where
-                    de.service = :service_name and
-                    (de.crn = :crn
-                         or de.noms_number = :noms_number )
-                 and (:start_date::date is null or de.created_at >= :start_date)
-                 and (:end_date::date is null or de.created_at <= :end_date) 
-             ) domain_events_metadata
       """.trimIndent(),
       MapSqlParameterSource()
         .addSarParameters(crn, nomsNumber, startDate, endDate)
