+++ src/main/resources/sar/template_hmpps-approved-premises-api.mustache
@@ -8,14 +8,10 @@
             <h4>Application</h4>
             <table class="data-table">
                 <tr>
-                    <td class="data-column-25">Application id</td>
-                    <td class="data-column-25">Offence id</td>
-                    <td class="data-column-25">Application user</td>
-                    <td class="data-column-25">Sentence type</td>
+                    <td class="data-column-50">Created by user</td>
+                    <td class="data-column-50">Sentence type</td>
                 </tr>
                 <tr>
-                    <td>{{ optionalValue id }}</td>
-                    <td>{{ optionalValue offence_id }}</td>
                     <td>{{ getUserLastName application_user_name }}</td>
                     <td>{{ convertCamelCase sentence_type }}</td>
                 </tr>
@@ -34,10 +30,9 @@
             </table>
             <table class="data-table">
                 <tr>
-                    <td class="data-column-25">Is withdrawn</td>
-                    <td class="data-column-25">Inmate in out status on submission</td>
-                    <td class="data-column-25">Created by</td>
-                    <td class="data-column-25">Notice type</td>
+                    <td class="data-column-30">Is withdrawn</td>
+                    <td class="data-column-30">Inmate in out status on submission</td>
+                    <td class="data-column-30">Notice type</td>
                 </tr>
                 <tr>
                     <td>{{ convertBoolean is_withdrawn }}</td>
@@ -47,29 +42,25 @@
             </table>
             <table class="data-table">
                 <tr>
-                    <td class="data-column-25">Arrival date</td>
-                    <td class="data-column-25">Target location</td>
-                    <td class="data-column-25">Conviction id</td>
-                    <td class="data-column-25">Case manager name</td>
+                    <td class="data-column-30">Arrival date</td>
+                    <td class="data-column-30">Target location</td>
+                    <td class="data-column-30">Case manager name</td>
                 </tr>
                 <tr>
                     <td>{{ optionalValue (formatDate arrival_date) }}</td>
                     <td>{{ optionalValue target_location }}</td>
-                    <td>{{ optionalValue conviction_id }}</td>
                     <td>{{ optionalValue case_manager_name }}</td>
                 </tr>
             </table>
             <table class="data-table">
                 <tr>
-                    <td class="data-column-25">Is emergency application</td>
-                    <td class="data-column-25">Release type</td>
-                    <td class="data-column-25">Name</td>
-                    <td class="data-column-25">Case manager is not applicant</td>
+                    <td class="data-column-30">Is emergency application</td>
+                    <td class="data-column-30">Release type</td>
+                    <td class="data-column-30">Case manager is not applicant</td>
                 </tr>
                 <tr>
                     <td>{{ convertBoolean is_emergency_application }}</td>
                     <td>{{ optionalValue release_type }}</td>
-                    <td>{{ optionalValue name }}</td>
                     <td>{{ convertBoolean case_manager_is_not_applicant }}</td>
                 </tr>
             </table>
@@ -89,12 +80,12 @@
             </table>
             <table class="data-table">
                 <tr>
-                    <td class="data-column-25">Type of approved premises</td>
-                    <td class="data-column-25">Check your answers reviewed</td>
+                    <td class="data-column-50">Type of approved premises</td>
+                    <td class="data-column-50"></td>
                 </tr>
                 <tr>
-                    <td>{{ optionalValue data.type-of-ap.ap-type.type }}</td>
-                    <td>{{ convertBoolean data.check-your-answers.review.reviewed }}</td>
+                    <td>{{ optionalValue ap_type }}</td>
+                    <td></td>
                 </tr>
             </table>
 
@@ -326,14 +317,6 @@
                     <td>{{ optionalValue criteria_type }}</td>
                 </tr>
             </table>
-            <table class="data-table">
-                <tr>
-                    <td class="data-column-50">Is active</td>
-                </tr>
-                <tr>
-                    <td>{{ convertBoolean is_active }}</td>
-                </tr>
-            </table>
         {{/PlacementRequirementCriteria}}
         {{^PlacementRequirementCriteria}}
             <p>No Data Held</p>
@@ -430,14 +413,6 @@
                 {{^sufficient-information}}
                     <p>No Data Held</p>
                 {{/sufficient-information}}
-
-                <p><b>Check Your Answers</b></p>
-                {{#check-your-answers}}
-                    <table class="summary-list">{{#each .}}<tr><td class="data-column-80">{{ @key }}</td><td class="data-column-20">{{ optionalValue this }}</td></tr>{{/each}}</table><br/>
-                {{/check-your-answers}}
-                {{^check-your-answers}}
-                    <p>No Data Held</p>
-                {{/check-your-answers}}
             {{/document}}
             {{^document}}
                 <p>No Data Held</p>
@@ -603,14 +578,10 @@
             </table>
             <table class="data-table">
                 <tr>
-                    <td class="data-column-25">Canonical arrival date</td>
-                    <td class="data-column-25">Canonical departure date</td>
-                    <td class="data-column-25">Expected arrival date</td>
-                    <td class="data-column-25">Expected departure date</td>
+                    <td class="data-column-50">Expected arrival date</td>
+                    <td class="data-column-50">Expected departure date</td>
                 </tr>
                 <tr>
-                    <td>{{ optionalValue (formatDate canonical_arrival_date) }}</td>
-                    <td>{{ optionalValue (formatDate canonical_departure_date) }}</td>
                     <td>{{ optionalValue (formatDate expected_arrival_date) }}</td>
                     <td>{{ optionalValue (formatDate expected_departure_date) }}</td>
                 </tr>
@@ -970,7 +941,6 @@
             {{/if}}
 
             {{#if (eq type "APPROVED_PREMISES_PLACEMENT_APPLICATION_WITHDRAWN")}}
-                <h4>Approved Premises Placement Application Withdrawn</h4>
                 <table class="data-table">
                     <tr>
                         <td class="data-column-30">Expected arrival</td>
