+++ src/test/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/cas2/integration/sar/Cas2SubjectAccessRequestServiceTest.kt
@@ -62,8 +62,7 @@ class Cas2SubjectAccessRequestServiceTest : Cas2SarTestBase() {
       "Assessments": [],
       "StatusUpdates": [],
       "StatusUpdateDetails": [],
-      "DomainEvents":  [],
-      "DomainEventsMetadata": []
+      "DomainEvents":  []
    }
     """.trimIndent()
     assertJsonEquals(expectedJson, result)
@@ -93,8 +92,7 @@ class Cas2SubjectAccessRequestServiceTest : Cas2SarTestBase() {
       "Assessments": [${cas2AssessmentsJson(assessment)}],
       "StatusUpdates": [],
       "StatusUpdateDetails": [],
-      "DomainEvents":  [],
-      "DomainEventsMetadata": []
+      "DomainEvents":  []
    }
     """.trimIndent()
     assertJsonEquals(expectedJson, result)
@@ -126,8 +124,7 @@ class Cas2SubjectAccessRequestServiceTest : Cas2SarTestBase() {
       "Assessments": [${cas2AssessmentsJson(assessment)}],
       "StatusUpdates": [],
       "StatusUpdateDetails": [],
-      "DomainEvents":  [],
-      "DomainEventsMetadata": []
+      "DomainEvents":  []
 
    }
     """.trimIndent()
@@ -162,8 +159,7 @@ class Cas2SubjectAccessRequestServiceTest : Cas2SarTestBase() {
       "Assessments": [${cas2AssessmentsJson(assessment)}],
       "StatusUpdates": [${cas2StatusUpdatesJson(statusUpdate)}],
       "StatusUpdateDetails": [${cas2StatusUpdateDetails(statusUpdateDetail)}],
-      "DomainEvents":  [],
-      "DomainEventsMetadata": []
+      "DomainEvents":  []
       
    }
     """.trimIndent()
@@ -205,8 +201,7 @@ class Cas2SubjectAccessRequestServiceTest : Cas2SarTestBase() {
       "Assessments": [${cas2AssessmentsJson(assessment)}],
       "StatusUpdates": [${cas2StatusUpdatesJson(statusUpdate)}],
       "StatusUpdateDetails": [${cas2StatusUpdateDetails(statusUpdateDetail)}],
-      "DomainEvents": [${domainEventJson(domainEvent, null, embeddedBooking = false)}],
-      "DomainEventsMetadata": [${domainEventsMetadataJson(domainEvent)}]
+      "DomainEvents": [${domainEventJson(domainEvent, null, embeddedBooking = false)}]
    }
     """.trimIndent()
     assertJsonEquals(expectedJson, result)
