+++ src/test/resources/sar/cas3-expected-report.html
@@ -347,8 +347,8 @@
     <td class="data-column-50">Multi-Agency Public Protection Arrangements status</td>
   </tr>
   <tr>
-    <td>No Data Held</td>
-    <td>NotFound</td>
+    <td>{lastUpdated&#x3D;[2023, 6, 26], level&#x3D;CAT A/LEVEL 1}</td>
+    <td>Retrieved</td>
   </tr>
 </table>
 <table class="data-table">
@@ -371,10 +371,10 @@
     <td class="data-column-25">Risk to Known Adult</td>
   </tr>
   <tr>
-    <td>No Data Held</td>
-    <td>No Data Held</td>
-    <td>No Data Held</td>
-    <td>No Data Held</td>
+    <td>Medium</td>
+    <td>Medium</td>
+    <td>Medium</td>
+    <td>Medium</td>
   </tr>
 </table>
 <table class="data-table">
@@ -384,9 +384,9 @@
     <td class="data-column-25">Risk of Serious Harm last updated</td>
   </tr>
   <tr>
-    <td>No Data Held</td>
-    <td>NotFound</td>
-    <td>No Data Held</td>
+    <td>Medium</td>
+    <td>Retrieved</td>
+    <td>15 May 2026</td>
   </tr>
 </table>
 <table class="data-table">
@@ -395,8 +395,8 @@
     <td class="data-column-50">Flags status</td>
   </tr>
   <tr>
-    <td>No Data Held</td>
-    <td>NotFound</td>
+    <td>Risk to Children, Risk to Known Adult, Risk to Prisoner, Risk to Public, Risk to Staff</td>
+    <td>Retrieved</td>
   </tr>
 </table>
